Added fix for exporting special chars in RTF like éóù... reported on the mailinglist:
hello all

There is a problem with the "french accentuated letters" with the
export  format harvard rtf. This problem is new with the 2.1 version.
When I change the new file havard.begin.layout by the older the problem
disappear. I don't unserstand where is the bug because the two files are
so different.
     thanks
    Martine Chazelas