Added highlighting in preview area for search text (based on patch 3121914 by Maximilian Lengsfeld).
The "transport" of the words to search has been modified to conform to the infrastructure established by patch 3117881.
SearchExpressionKeyParser is NOT used, but the SearchTextListener interface directly passing the keywords.
The issue with this approach is that the search format "field=keyword" does not work. "field=keyword" is tried to be highlighted
and not "keyword" everywhere (as suggested by the patch).

Only the preview panel is highlighted, other layouts (export, find duplicates, ...) are not affected.