Fixed bug #1153 (and duplicate #1150).
I have fixed the bugs 1153 and 1150 (duplicates) that I originally
introduced a long time ago. Retained (read) handling of legacy setting
containing the problematic '@'. To get rid of the setting containing
'@', users must manually remove this line from their database file. The
latter is not a bug but an intended feature.