Update of JabRef_it.properties
Some of the translations might be inaccurate because of the difficulty of translating in Italian out-of-context sentences, in particular when no hint is available about gender and number of subject/object. E.g. the string "none" may be translated as "nessuna" or "nessun" depending of the gender of the item it refers to: "nessuna" if referring to an entry but "nessun" if referring to a group/number/title/author, or even "nessuno" if referring to space/style. In those cases I often do an educated guess on the basis of the French and German translations.