Fixes #1482: Correct number of matched entries is displayed for refining subgroups
The issue was that, for the number of hits, JabRef only checks if the
group taken separately matches the entry. That is, it is completely
ignored that the group sits in a tree and might be set up to refine the
search of the parent. Taking the example given in #1482, 577 entries
satisfy the condition `articlé original` but only 38 are also matched by
`articlé`.